+++ app/helpers/follows_helper.rb
@@ -1,0 +1,2 @@
+module FollowsHelper
+end
