--- app/helpers/relationships_helper.rb
@@ -1,2 +1,0 @@
-module RelationshipsHelper
-end
