+++ app/models/follow.rb
@@ -1,0 +1,4 @@
+class Follow < ApplicationRecord
+  belongs_to :follower, class_name: "User"
+  belongs_to :followed, class_name: "User"
+end
