--- app/models/relationship.rb
@@ -1,4 +1,0 @@
-class Relationship < ApplicationRecord
-    belongs_to :follower, class_name: "User"
-    belongs_to :followed, class_name: "User"
-end
