+++ db/migrate/20221018202539_create_follows.rb
@@ -1,6 +1,6 @@
-class CreateRelationships < ActiveRecord::Migration[7.0]
+class CreateFollows < ActiveRecord::Migration[7.0]
   def change
-    create_table :relationships do |t|
+    create_table :follows do |t|
       t.integer :follower_id
       t.integer :followed_id
